# api/openapi/risk-context.yaml
openapi: 3.0.3
info:
  title: Risk Context Service API
  version: 1.0.0
  description: Transaction risk assessment APIs.
servers:
  - url: https://api.sandbox.openfinance.ae
paths:
  /api/v1/risk/assess:
    post:
      summary: Assess transaction risk
      operationId: assessRisk
      description: |-
        Decides once per transactionId. A retry with the same amount and
        currency returns the stored decision; the same transactionId with a
        different amount or currency is a 409.
      security:
        - bearerAuth: []
          dpopAuth: []
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/DPoP'
        - $ref: '#/components/parameters/InteractionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskAssessmentRequest'
      responses:
        '201':
          description: Risk assessment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskAssessmentResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Transaction already assessed with a different amount or currency
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/risk/assessments/{transactionId}:
    get:
      summary: Find risk assessment by transaction id
      operationId: findRiskAssessment
      security:
        - bearerAuth: []
          dpopAuth: []
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/DPoP'
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/TransactionId'
      responses:
        '200':
          description: Risk assessment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskAssessmentResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No assessment for this transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    dpopAuth:
      type: http
      scheme: bearer
      bearerFormat: DPoP
  parameters:
    Authorization:
      name: Authorization
      in: header
      required: true
      schema:
        type: string
    DPoP:
      name: DPoP
      in: header
      required: true
      schema:
        type: string
    InteractionId:
      name: X-FAPI-Interaction-ID
      in: header
      required: true
      schema:
        type: string
    TransactionId:
      name: transactionId
      in: path
      required: true
      schema:
        type: string
  schemas:
    RiskAssessmentRequest:
      type: object
      required: [transactionId, amount, currency, velocityScore]
      properties:
        transactionId:
          type: string
        amount:
          type: number
          format: double
        currency:
          type: string
        highRiskCountry:
          type: boolean
        velocityScore:
          type: integer
          minimum: 0
          maximum: 100
    RiskAssessmentResponse:
      type: object
      required: [assessmentId, transactionId, score, decision, reasons, assessedAt]
      properties:
        assessmentId:
          type: string
        transactionId:
          type: string
        score:
          type: integer
        decision:
          type: string
          enum: [ALLOW, REVIEW, BLOCK]
        reasons:
          type: array
          items:
            type: string
        assessedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [code, message, interactionId, timestamp]
      properties:
        code:
          type: string
        message:
          type: string
        interactionId:
          type: string
        timestamp:
          type: string
          format: date-time